migration-src/application/test.py
#!/usr/bin/env python
# -*- coding: utf-8 -*-
"""
Test endpoint for verifying Python 3.11 migration
将来的なテストコードを追加する場所として使用
"""

import os
import sys
from flask import jsonify, request
from google.cloud import ndb
import datetime


def test_route():
    """
    Flask route function for test endpoint
    Python 3.11 マイグレーションの動作確認用
    """
    tests_results = []

    # Test 1: Basic Python version check
    tests_results.append({
        "test": "Python Version",
        "result": f"{sys.version}",
        "status": "OK"
    })

    # Test 2: Environment check
    tests_results.append({
        "test": "Runtime Environment",
        "result": f"GAE_ENV={os.getenv('GAE_ENV', 'Not set')}",
        "status": "OK"
    })

    # Test 3: NDB Client availability
    try:
        client = ndb.Client()
        tests_results.append({
            "test": "NDB Client",
            "result": "Cloud NDB client initialized",
            "status": "OK"
        })
    except Exception as e:
        tests_results.append({
            "test": "NDB Client",
            "result": f"Error: {str(e)}",
            "status": "ERROR"
        })

    # Test 4: Request info
    tests_results.append({
        "test": "Request Method",
        "result": request.method,
        "status": "OK"
    })

    # Test 5: Datetime
    tests_results.append({
        "test": "Current Time (UTC)",
        "result": datetime.datetime.utcnow().isoformat(),
        "status": "OK"
    })

    # Summary
    total_tests = len(tests_results)
    passed_tests = sum(1 for t in tests_results if t["status"] == "OK")

    response = {
        "service": "test-service",
        "runtime": "Python 3.11",
        "endpoint": "/test",
        "summary": {
            "total": total_tests,
            "passed": passed_tests,
            "failed": total_tests - passed_tests
        },
        "tests": tests_results,
        "message": "Python 3.11 migration test endpoint - OK"
    }

    return jsonify(response)


# 将来的なテスト関数を追加する場所
# 例:
# def test_datastore_connection():
#     """Test Cloud Datastore connection"""
#     pass
#
# def test_model_operations():
#     """Test NDB model CRUD operations"""
#     pass
