migration-src/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
google-cloud-datastore==2.19.0
google-cloud-storage==2.14.0
google-cloud-ndb==2.3.0
google-cloud-tasks==2.16.0
requests==2.31.0
